refactor: make tests function with the new environment
