+++ react-app/src/testing/tests/Api.test.tsx
@@ -1,0 +1,8 @@
+import { describe, expect, it } from "vitest";
+
+describe("Testing Api", () => {
+  it("calling with a pokemin name", async () => {
+    // const response = await getPokemonDataByName("");
+    // expect(response.data.length).toBe(20);
+  });
+});
