+++ react-app/src/types.ts
@@ -1,31 +1,6 @@
-export type PokemonCardData = {
-  name: string;
-  key: number;
-  imgUrl: string | null;
-  descr: string;
-};
-
-export type PokemonDataResponse = {
-  data: (PokemonCardData | undefined)[];
-  next?: string | null;
-  prev?: string | null;
-};
-
-export interface PokemonElementProps {
-  pokemons: PokemonCardData[];
-}
-
-export interface PokemonElementState extends PokemonElementProps {
-  loading: boolean;
-}
-
-export type SearchBarProps = {
-  search: (name: string) => void;
-};
-
-export type SearchBarData = {
+export interface SearchBarData {
   inputValue: string;
-};
+}
 
 export interface QueryParams {
   search?: string;
