+++ react-app/vite.config.ts
@@ -11,6 +11,6 @@ export default defineConfig({
   test: {
     globals: true,
     environment: "jsdom",
-    setupFiles: "./setupTests.ts",
+    setupFiles: "./src/testing/setupTests.ts",
   },
 });
